# apps/ai/openapi.json
{
  "components": {
    "schemas": {
      "CoverageAnalyzeAcceptedResponse": {
        "properties": {
          "status": {
            "const": "accepted",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "CoverageAnalyzeAcceptedResponse",
        "type": "object"
      },
      "CoverageAnalyzeRequest": {
        "properties": {
          "case_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Case Id",
            "type": "string"
          },
          "claim_summary": {
            "title": "Claim Summary",
            "type": "string"
          },
          "document_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Document Ids",
            "type": "array"
          }
        },
        "required": [
          "case_id",
          "document_ids",
          "claim_summary"
        ],
        "title": "CoverageAnalyzeRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MatterCitation": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "page_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page End"
          },
          "page_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Start"
          },
          "section_label": {
            "title": "Section Label",
            "type": "string"
          },
          "similarity": {
            "title": "Similarity",
            "type": "number"
          }
        },
        "required": [
          "id",
          "section_label",
          "page_start",
          "page_end",
          "content",
          "similarity",
          "document_id"
        ],
        "title": "MatterCitation",
        "type": "object"
      },
      "PolicyCitation": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "page_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page End"
          },
          "page_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Start"
          },
          "section_label": {
            "title": "Section Label",
            "type": "string"
          },
          "similarity": {
            "title": "Similarity",
            "type": "number"
          }
        },
        "required": [
          "id",
          "section_label",
          "page_start",
          "page_end",
          "content",
          "similarity"
        ],
        "title": "PolicyCitation",
        "type": "object"
      },
      "PolicyQAMatterRequest": {
        "properties": {
          "document_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Document Ids",
            "type": "array"
          },
          "question": {
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "document_ids",
          "question"
        ],
        "title": "PolicyQAMatterRequest",
        "type": "object"
      },
      "PolicyQAMatterResponse": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/MatterCitation"
            },
            "title": "Citations",
            "type": "array"
          },
          "injection_warnings": {
            "items": {
              "type": "string"
            },
            "title": "Injection Warnings",
            "type": "array"
          },
          "refused": {
            "title": "Refused",
            "type": "boolean"
          },
          "retrieved_chunks": {
            "items": {
              "$ref": "#/components/schemas/MatterCitation"
            },
            "title": "Retrieved Chunks",
            "type": "array"
          },
          "top_similarity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Similarity"
          }
        },
        "required": [
          "answer",
          "citations",
          "retrieved_chunks",
          "refused",
          "top_similarity"
        ],
        "title": "PolicyQAMatterResponse",
        "type": "object"
      },
      "PolicyQARequest": {
        "properties": {
          "document_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Document Id",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "question"
        ],
        "title": "PolicyQARequest",
        "type": "object"
      },
      "PolicyQAResponse": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/PolicyCitation"
            },
            "title": "Citations",
            "type": "array"
          },
          "injection_warnings": {
            "items": {
              "type": "string"
            },
            "title": "Injection Warnings",
            "type": "array"
          },
          "refused": {
            "title": "Refused",
            "type": "boolean"
          },
          "retrieved_chunks": {
            "items": {
              "$ref": "#/components/schemas/PolicyCitation"
            },
            "title": "Retrieved Chunks",
            "type": "array"
          },
          "top_similarity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Similarity"
          }
        },
        "required": [
          "answer",
          "citations",
          "retrieved_chunks",
          "refused",
          "top_similarity"
        ],
        "title": "PolicyQAResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/coverage/analyze": {
      "post": {
        "operationId": "coverage_analyze_coverage_analyze_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoverageAnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverageAnalyzeAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Coverage Analyze",
        "tags": [
          "coverage"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/qa/answer": {
      "post": {
        "operationId": "qa_answer_qa_answer_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyQARequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyQAResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Qa Answer",
        "tags": [
          "qa"
        ]
      }
    },
    "/qa/matter": {
      "post": {
        "operationId": "qa_matter_qa_matter_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyQAMatterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyQAMatterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Qa Matter",
        "tags": [
          "qa"
        ]
      }
    }
  }
}